+++ main/api/origins.ts
@@ -8,30 +8,12 @@ import store from '../store'
 
 const dev = process.env.NODE_ENV === 'development'
 
-function invalidOrigin (origin: string) {
+export function invalidOrigin (origin: string) {
   return origin !== origin.replace(/[^0-9a-z/:.[\]-]/gi, '')
 }
 
 export function getOriginId(originName: string) {
   return uuidv5(originName, uuidv5.DNS) as UUID<Origin>
-}
-
-function addPermissionRequest (address: Address, origin: string) {
-  return new Promise((resolve, reject) => {
-    const handlerId = uuidv5(origin, uuidv5.DNS)
-    const request: AccessRequest = { payload: undefined as any, handlerId, type: 'access', origin: handlerId, account: address }
-
-    accounts.addRequest(request, () => {
-      const permissions = store('main.permissions', address) || {}
-      const perms = Object.keys(permissions).map(id => permissions[id])
-      const permIndex = perms.map(p => p.origin).indexOf(origin)
-      if (perms[permIndex] && perms[permIndex].provider) {
-        resolve(true)
-      } else {
-        reject(new Error('Origin does not have provider permissions'))
-      }
-    })
-  })
 }
 
 export function updateOrigin (payload: JSONRPCRequestPayload, originName?: string, connectionMessage = false): RPCRequestPayload {
@@ -79,28 +61,5 @@ export function isFrameExtension (req: IncomingMessage) {
     return query.identity === 'frame-extension'
   } else {
     return false
-  }
-}
-
-export async function isTrusted (origin?: string) {
-  if (!origin || origin === 'null') origin = 'Unknown' // Permission granted to unknown origins only persist until the Frame is closed, they are not permanent
-  if (invalidOrigin(origin)) return false
-  if (origin === 'frame-extension') return true
-  const account = accounts.current()
-  if (!account) return
-  const address = account.address
-  if (!address) return
-  const permissions = store('main.permissions', address) || {}
-  const perms = Object.keys(permissions).map(id => permissions[id])
-  const permIndex = perms.map(p => p.origin).indexOf(origin)
-  if (permIndex === -1) {
-    try {
-      return await addPermissionRequest(address, origin)
-    } catch (e) {
-      log.error(e)
-      return false
-    }
-  } else {
-    return perms[permIndex] && perms[permIndex].provider
   }
 }
