+++ main/api/ws.ts
@@ -7,7 +7,8 @@ import provider from '../provider'
 import accounts from '../accounts'
 import windows from '../windows'
 
-import { updateOrigin, isTrusted, isFrameExtension, getOriginId } from './origins'
+import { updateOrigin, isFrameExtension, getOriginId } from './origins'
+import { isTrusted } from './permissions'
 import validPayload from './validPayload'
 import protectedMethods from './protectedMethods'
 import { IncomingMessage, Server } from 'http'
